requirements.txt
Django
psycopg[binary]
pillow
python-dotenv
google-generativeai
